+++ src/app/app.component.html
@@ -1,5 +1,5 @@
 <div class="main">
-  <h1>ToDo List</h1>
+  <h1>{{title}}</h1>
   <router-outlet></router-outlet>
 </div>
