+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component, VERSION } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
@@ -6,41 +6,5 @@ import { Component, VERSION } from '@angular/core';
   styleUrls: [ './app.component.css' ]
 })
 export class AppComponent  {
-  name = 'TodoList-Angular' + VERSION.major;
-  title = 'ToDoList';
-
-  filter: 'all' | 'active' | 'done' = 'all';
-
-  allItems = [
-    { description: 'eat', done: true },
-    { description: 'sleep', done: false },
-    { description: 'play', done: false },
-    { description: 'laugh', done: false },
-  ];
-
-  get items() {
-    if (this.filter === 'all') {
-      return this.allItems;
-    }
-    return this.allItems.filter(item => this.filter === 'done' ? item.done : !item.done);
-  }
-
-  countItems(filter: string) {
-    if (filter === 'all') {
-      return this.allItems.length;
-    }
-    return this.allItems.filter(item => filter === 'done' ? item.done : !item.done).length;
-  }
-
-  addItem(description: string){
-    // unshift (to top), push (to end)
-    this.allItems.unshift({
-      description,
-      done: false
-    });
-  }
-  
-  remove(item) {
-    this.allItems.splice(this.allItems.indexOf(item), 1);
-  }
+  title = 'ToDo List';
 }
